requirements.txt
sanic==18.12.0
docker_py==1.10.6
